requirements.txt
asgiref==3.8.1
# backports.zoneinfo==0.2.1
black==24.4.2
click==8.1.7
Django==4.2.11
djangorestframework==3.15.1
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
sqlparse==0.5.0
tomli==2.0.1
typing-extensions==4.11.0
